pyproject.toml
[project]
name = "brd4-kan"
version = "0.1.0"
description = "BRD4-KAN: reproducible symbolic QSAR and virtual screening pipeline using advanced Kolmogorov-Arnold Networks."
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = { text = "MIT" }
authors = [{ name = "BRD4-KAN authors" }]

dependencies = [
    # Cheminformatics
    "rdkit==2024.3.5",
    "chembl_structure_pipeline==1.2.2",
    "mordredcommunity==2.0.6",
    "datamol==0.12.4",
    # Classical ML / baselines
    "scikit-learn==1.5.2",
    "xgboost==2.1.1",
    "chemprop==2.0.4",
    # KAN stack
    "efficient-kan==0.1.0",
    "pykan==0.2.8",
    # Torch
    "torch==2.4.1",
    # Experiment infra
    "optuna==4.0.0",
    "mlflow==2.17.0",
    "dvc[all]==3.55.2",
    "hydra-core==1.3.2",
    "typer==0.12.5",
    "pydantic==2.9.2",
    # Figures
    "matplotlib==3.9.2",
    "cairosvg==2.7.1",
    "svgutils==0.3.4",
    # Utilities
    "numpy==1.26.4",
    "pandas==2.2.3",
    "pyyaml==6.0.2",
    "python-dotenv==1.0.1",
    "joblib==1.4.2",
    "tqdm==4.66.5",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-cov==5.0.0",
    "ruff==0.6.9",
    "mypy==1.11.2",
    "pre-commit==3.8.0",
    "types-PyYAML==6.0.12.20240917",
]

[project.scripts]
brd4kan = "brd4kan.cli:app"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/brd4kan"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "NPY", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
ignore_missing_imports = true
packages = ["brd4kan"]
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra --strict-markers --cov=brd4kan --cov-report=term-missing --cov-fail-under=85"
markers = [
    "slow: marks tests as slow (end-to-end smoke, deselect with '-m \"not slow\"')",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src/brd4kan"]
branch = true
